cache known http headers
Every time a request comes, new string is allocated for the http
field. However this can be avoided if the http field is one of the
known http header name. This commit exposes `knownHttpHeaders`
JsArray that stores pre-allocated string which can be used to save
inside `rawHeaders` or `rawTrailers`. However there is a cost to
search for the field name in the lookup in C++ land. I tried multiple
implementation to get this not regress and here is implementation
I came up with.

There is a `unordered_map` that stores hashcode of field name to
index. Index denotes the entry of JsArray where field name is stored.
We look the field name in the map by first calculating the hash code.
If we get an entry, we pass the index in `headers` which is coverted
to equivalent headerName in Jsland.